analyst/pyproject.toml
[project]
name = "analyst"
version = "0.1.0"
description = "Agentic AI using CrewAI"
authors = [
    { name = "Neha Sureddi", email = "nehasureddi@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "crewai-tools",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
    "tabulate>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
run_crew = "analyst.main:run"
train = "analyst.main:train"
replay = "analyst.main:replay"
test = "analyst.main:test"

[tool.crewai]
type = "crew"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
